[model] Add prototype of class for decomposing a model in several submodels
The concept of dividing a model in several submodels is present in several
algorithms implemented in iDynTree. What we have usually done is to
split the iCub model along the six axis F/T sensor joints, to exploit
the sensor measurement for external forces estimation and inertial parameter
identification. Support for this has been rather bad implemented in the past,
but now we can work on streamline it.

Every submodel is associated with a Traversal object, in this
way a given algorithm can run on both the full model or
on a submodel.