[sensors] Add Sensor postfix to Accelerometer and Gyroscope class names
This is necessary because the ACCELEROMETER enum value was mapped in matlab
to the ACCELEROMETER.m file, while the Accelerometer class was mapped to the
Accelerometer.m file. There was not problem in Linux, but in OS X and Windows
the filesystem are case insensitive, and so this clash was causing problems even
when just cloning the repo.

cc @naveenoid