[yarp] Fix the angular-linear serialization used in TorqueEstimation
Fix https://github.com/robotology/idyntree/issues/123

For some reason, the angular-linear serialization was used just in
the KDLtoYarp_twist and KDLtoYarp_wrench functions, and in the TorqueEstimationTree::buildAb_contacts
and TorqueEstimationTree::store_contacts_results methods. Switch to use
linear-angular serialization for consistency with the rest of the code.

This should not affect the input-output behaviour of the TorqueEstimationTree class.