[Sensors] Create function to extract a reduced model and sensors list
Furthermore, fix various issues around in the code.

The spreading of code related to this in the library is suboptimal,
and mostly due to the fact that model functions can't depend on sensors data structure.
The definite fix for this is to merge the model and sensors libraries, but
this require that first we solve this issue: https://github.com/robotology/idyntree/issues/132 .