[java] Rename u public attribute in ArticulatedBodyInertia buffers
In preparation of the Java bindings generation, rename u in uu.
This is necessary because every public attribute is going to be wrapped
 in Java with a couple of automatically generated set and get, and
the presence in the same class of U and u was generatign two identically
getU() methods.

This change brecks the compilation of Matlab bindings, that will need to
be regenerated and commited before merging the feature/java branch in master.