Added visualization library
Add an optional depedency on the Irrlicht graphical engine,
choosen because it has a limited amount of dependency and it is
available on apt-get , homebrew and in binary form for Windows.

The dependency can be enabled with the IDYNTREE_USES_IRRLICHT variable.

The Visualizer class is currently quite limited, but the `idyntree-model-view`
has been added as a tool to showcase its use (for now it just visualize statically
a model in its default position).