Adding SparseMatrix and Triplet(s) classes
- Triplets is a simple container of Triplet classes which exposes
utility methods to set more than one Triplet at the time,
e.g. setSubMatrix.
- Triplet describes an element of a matrix, i.e. its coordinates (row and column)
and its value.
- SparseMatrix is a new class which represents... guess what... sparse matrices.
Its main scope is to provide an iDynTree-based container for Sparse matrices,
with the idea that they will be mapped to Eigen-based SparseMatrix classes to
perform any mathematical operation.
For this reason we implemented a Row Compressed Storage scheme (RCS).
Eigen compressed sparse matrices implement the same scheme, thus iDynTree matrices
can be used with Eigen Map.
Note that iDynTree SparseMatrix uses a RowMajor organization.
This has been done to be coherent with the Dense counterpart, but given the big difference
between dense and sparse matrices we may drop this peculiarity in the future.