Fix getAxis and setAxis implementation in Revolute and Prismatic Joint and documentation cleanup
The direction of the axis in a joint depends on which body
you consider parent and which body you consider child .
See "Minimal formulation of joint motion for biomechanisms",
Section 2.4 and Traversaro's PhD Thesis, Chapter 3.

This is a small change, but it is extremly important.
I hope that the additional documentation will help document it.