Added ColumnMajor storage order SparseMatrix
Now SparseMatrix is a template depending on the storage ordering (row-major or column-major)
All the methods are implemented in the .cpp file.
The library instantiate the two (only) possible SparseMatrix types, i.e. row and column major,
so the implementation of the methods should not be necessary, thus the choice to put it in the .cpp
Furthermore, the actual difference between the two template version is the implementation of some methods.
Template here is used as an alternative to hierarchy.