Added debug-only check code
Discussing with @pattacini and @S-Dafarra we decided to add a check in IpOpt (running only when the code is compiled in debug).
We now check that a callback function is called only once for a specified set of input parameters.
If this does not happen, the code is still correct, but we execute some code twice.
As it has been tested now, this does not happen, so the code does not need any change at this time.