Modified the L2Norm cost and the quadratic cost to allow correct evaluation of norm.
When translating a L2Norm cost into a quadratic cost, the constant part given by the square of the reference value can be omitted. On the other end, for applications when it is important to have the correct value of the 2-norm, this constant part can be computed and added to the cost.