Fix multiple calls to add**(Target|Constraint)
While the InverseKinematics library was designed to specify all the
targets and constraints **before** the  first call to solve,
apparently for some users it make sense to call again this method
between solve, to change some aspects of the constraints.

While this is not properly documented, this commit make sures
that if the user uses the methods in this way, everything works fine.