Add SolidShapes library that depends on Assimp and related inertia parameter estimation helpers
This commit adds a SolidShape helper library. This library is part of iDynTree, and is meant
to contain all the algorithms that use in some form the visual and collision geometries of the model,
and so they depend on the Assimp library to load meshes.

The first algorithm added to the new library is an helper function that provides rough estimates
of the inertial parameters (mass, first moments of mass, 3d inertia matrix elements) of a robot
given the total mass of the robot, and its collisions shapes. While the estimates provided are
quite rough, they can be quite useful at least to provide an expected order of magnitute of the parameters,
to normalize errors or as initial points of a nonlinear optimization procedure.