Remove generic never called methods for SparseMatrix
Some methods of SparseMatrix need to be implemented specifically for the
required ordering, and are never instantiated for an order different from
either iDynTree::RowMajor or iDynTree::ColumnMajor.
By removing them, we reduce the amount of useless code, and we avoid spurious warnings.

Fix https://github.com/robotology/idyntree/issues/605