Remove all uses of IDYNTREE_TREE_INCLUDE_DIRS global variable
The IDYNTREE_TREE_INCLUDE_DIRS global variable dates back to CMake 2.8.7,
when the "modern CMake" propagation of include usage requiremements via
target_include_directories was not available, and one of the main iDynTree
authors had quite a limited understanding of CMake in general.

It used to contain all the include directories of iDynTree, but not it is not
necessary. The only place in which it was still useful was:
* Passing the include directories to SWIG. For this, now all the idyntree targets
  are inspected and the appropriate include directories are extracted from the
  target-specific INTERFACE_INCLUDE_DIRECTORIES properties.
* Passing to tests the location of the testModels.h header that contains the location
  of test model files. This usage has been substitute with a new HEADER library idyntree-testmodels,
  that is linked in all tests that need to use the testModels.h header.