Support the use of dependencies made available via add_subdirectory
If iDynTree is included as part of a bigger project via add_subdirectory,
it is possible that some of its dependencies are included via add_subdirectory as well.
For this reason, this PR adds the logic to avoid searching via find_package for
a CMake package, if the relevant target are already defined.