Add definition of namespaced targets even for the build tree
It is useful to define an alias of the iDynTree libraries with the scope of the exported
library itself (i.e. iDynTree::<lib>) because (1) you can link against it with the exact same syntax
of an imported library, that is useful when iDynTree is used in a bigger project
via add_subdirectory and (2) because names having a double-colon (::) are
always treated as the name of either an alias or imported target. Any attempt
to use such a name for a different target type will result in an error.