Change compatibility of CMake config files from SameMajorVersion to AnyNewerVersion
Even if in the future major version can be breaking, for the main classes it would be 
relativly easy to have a codebase that is compatible with multiple iDynTree major versions. 
For this reason, I think it make sense to switch the CMake policy to `AnyNewerVersion` .
If any downstream user wants to use a different policy, it can always manually check inspect the `iDynTree_VERSION` 
CMake variable.