Remove additional project and cmake_minimum_required calls in nested directories (#638)
* Remove additional project and cmake_minimum_required calls in nested functions

The definition of a project in CMake is tipically used to indicate
a self-sufficient project that can be built on its own. This was not the case
for most of the use of project. Those calls just created confusion on the values
of all the PROJECT_ variables, so they are not really useful and they can be
simply removed.