Avoid to use the old GenerateExportHeader included with VTK 6
vtk 6.3 installs a GenerateExportHeader CMake module that shadows the official CMake module if find_package(VTK)
is invoked. The issue has been solved in vtk >= 7.1, see https://gitlab.kitware.com/vtk/vtk/-/merge_requests/1593 .
To ensure that the build works even if find_package(VTK) is invoked and VTK 6 is found, we manually include exactly the
GenerateExportHeader that is shipped with CMake, extracting the CMake install prefix from the CMAKE_COMMAND variable
This workaround can be removed once we will not support anymore platforms that ship with vtk 6.3 as main vtk version (Ubuntu 18.04).