Transform idyntree-yarp and idyntree-icub to header only libraries
The `idyntree-yarp` and `idyntree-icub` libraries are now header-only, and are always installed even if
`IDYNTREE_USES_YARP` or `IDYNTREE_USES_ICUB_MAIN` are set to `OFF`, to simplify deployment of the library.
The downstream libraries that want to use them need to find and link `YARP` and `ICUB` CMake packages on their own.